pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "looped-experiments"
version = "0.1.0"
authors = [{ name = "Slava Chaunin", email = "slava.chaunin@gmail.com" }]
readme = "README.md"
dependencies = [
    "torch",
    "numpy",
    "fastcore",
    "fastprogress",
    "matplotlib",
    "hydra-core",
]

[project.scripts]
launch-exps = "looped_experiments.main:run"

[project.urls]
Homepage = "https://github.com/ssslakter/tg-rag"
Issues = "https://github.com/ssslakter/tg-rag/issues"

[tool.setuptools]
packages = ["looped_experiments"]


